+++ package.json
@@ -36,7 +36,7 @@
     "i": "^0.3.7",
     "immer": "^10.0.2",
     "jsonwebtoken": "8.5.1",
-    "next": "^13.4.12",
+    "next": "^13.4.19",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-icons": "^4.10.1",
@@ -53,7 +53,7 @@
     "@types/node": "^20.4.5",
     "@types/react": "^18.2.0",
     "eslint": "8.8.0",
-    "eslint-config-next": "12.0.10",
+    "eslint-config-next": "^13.4.19",
     "eslint-config-prettier": "^8.3.0",
     "husky": "^7.0.4",
     "lint-staged": "^12.3.3",
